% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMtBroadPeak.R
\name{check_identical_header}
\alias{check_identical_header}
\title{helper function to check the bam file header}
\usage{
check_identical_header(bams)
}
\arguments{
\item{bams}{Bam file of treatments and controls}
}
\value{
seqlengths
}
\description{
helper function to check the bam file header
}
