% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeaks.R
\name{plotPeaks}
\alias{plotPeaks}
\title{Plot tool for called peaks}
\usage{
plotPeaks(res, seqname)
}
\arguments{
\item{res}{output of \link{HMMtBroadPeak}.}

\item{seqname}{the chromosome to plot.}
}
\value{
an ggplot object
}
\description{
Plot the log2 transformed signal along a given chromosome with the peaks.
}
